Add patches for `pytest.mark.parametrize`
PyCharm provides skeletons for autocompleting various Pytest functions; I
assume it does this because a lot of Pytest's code is dynamically imported and
therefore difficult for PyCharm to inspect. One such function is `parametrize`,
which is used to define parametrised tests. The signature of this function has
been extended since PyCharm last updated their skeleton. This commit adds a
patch to amend the relevant file. It also adds a pre-patch that normalises
trailing whitespace in the same file, for ease of editing.