build.sh: Fail on errors
This is particularly useful when a build is run as a non-privileged user after
the previous build was run as 'root'. In this scenario, the results directory
will be owned by 'root', so the 'rf -rf' will fail. If the script continues,
further issues will occur and the RPM will not be built. This commit avoids
that scenario by failing early.